server/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
bcrypt==4.3.0
pydantic==2.11.3
psycopg2-binary==2.9.10
